+++ views/js/validation.js
@@ -63,10 +63,6 @@ $(document).ready(function () {
     function validatePaymentMethod() {
         var $paymentMethodForm = $(this).closest('.payment-method');
 
-        var $isPaymentEnabled = $paymentMethodForm.find('select[name^="MOLLIE_METHOD_ENABLED"]');
-        if ($isPaymentEnabled.val() === "0") {
-            return;
-        }
         var $description = $paymentMethodForm.find('input[name^="MOLLIE_METHOD_DESCRIPTION"]');
         if (!/\S/.test($description.val())) {
             event.preventDefault();